reduce surface area of audio crates II
