+++ crates/unaudiospatial-core/Cargo.toml
@@ -13,3 +13,5 @@ workspace = true
 [dependencies]
 unspatial-core = { workspace = true }
 bevy = { workspace = true }
+bevy_asset_loader = { workspace = true }
+bevy_seedling = { workspace = true }
