+++ crates/unaudiospatial-core/src/components.rs
@@ -1,5 +1,19 @@
 use bevy::prelude::*;
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AudioCategory {
+    Effects,
+    VoiceChat,
+    Master,
+}
+
+#[derive(Component, Debug, Clone)]
+pub struct FlatAudio {
+    pub sound_file: String,
+    pub volume_multiplier: f32,
+    pub category: AudioCategory,
+}
+
 #[derive(Component, Debug, Clone)]
 pub struct SpatialAudioInstance {
     pub sound_file: String,
