+++ crates/unmapload-core/Cargo.toml
@@ -12,6 +12,5 @@ workspace = true
 
 [dependencies]
 bevy = { workspace = true }
-bevy_seedling = { workspace = true }
 bevy_asset_loader = { workspace = true }
 unbehavior-core = { workspace = true }
